+++ keep-ui/entities/workflows/index.ts
@@ -1,1 +1,4 @@
 export { useWorkflowActions } from "./model/useWorkflowActions";
+export { useWorkflowStore } from "./model/workflow-store";
+
+export type * from "./model/types";
