+++ keep-ui/features/workflows/builder/ui/NodeMenu.tsx
@@ -2,9 +2,9 @@ import { Menu, Transition } from "@headlessui/react";
 import { Fragment } from "react";
 import { CiSquareChevDown } from "react-icons/ci";
 import { TrashIcon } from "@heroicons/react/24/outline";
-import useStore from "./builder-store";
+import { useWorkflowStore } from "@/entities/workflows";
 import { IoMdSettings } from "react-icons/io";
-import { FlowNode } from "@/app/(keep)/workflows/builder/types";
+import { FlowNode } from "@/entities/workflows/model/types";
 
 export default function NodeMenu({
   data,
@@ -20,7 +20,8 @@ export default function NodeMenu({
     data?.type?.includes("empty") ||
     id?.includes("end") ||
     id?.includes("start");
-  const { deleteNodes, setSelectedNode, setStepEditorOpenForNode } = useStore();
+  const { deleteNodes, setSelectedNode, setStepEditorOpenForNode } =
+    useWorkflowStore();
 
   return (
     <>
