+++ keep-ui/features/workflows/builder/ui/__tests__/ReactFlowBuilder.test.tsx
@@ -2,7 +2,7 @@ import { act, render, renderHook } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import ReactFlowBuilder from "../ReactFlowBuilder";
 import { ReactFlowProvider } from "@xyflow/react";
-import useStore from "../builder-store";
+import { useWorkflowStore } from "@/entities/workflows";
 
 // Mock the hooks and components
 jest.mock("../ToolBox", () => ({
@@ -33,7 +33,7 @@ describe("ReactFlowBuilder", () => {
   });
 
   it("renders successfully", () => {
-    const { result } = renderHook(() => useStore());
+    const { result } = renderHook(() => useWorkflowStore());
 
     act(() => {
       result.current.setDefinition({
