+++ keep-ui/features/workflows/edit-metadata/index.ts
@@ -1,0 +1,1 @@
+export { WorkflowMetadataModal } from "./ui/workflow-metadata-modal";
